Move series-manipulation to util
