+++ documentation/ChapterArticle.json
@@ -133,5 +133,5 @@
     "value" : "1234"
   } ],
   "subjects" : [ "http://example.org/subject/123" ],
-  "modelVersion" : "0.13.0"
+  "modelVersion" : "0.13.1"
 }
