+++ documentation/JournalCorrigendum.json
@@ -137,5 +137,5 @@
     "value" : "1234"
   } ],
   "subjects" : [ "http://example.org/subject/123" ],
-  "modelVersion" : "0.13.0"
+  "modelVersion" : "0.13.1"
 }
